+++ src/api/coinmarketcap.js
@@ -2,6 +2,35 @@ import axios from 'axios'
 
 const CURRENCY = ['aud', 'brl', 'cad', 'chf', 'clp', 'cny', 'czk', 'dkk', 'eur', 'gbp', 'hkd', 'huf', 'idr', 'ils', 'inr', 'jpy', 'krw', 'mxn', 'myr', 'nok', 'nzd', 'php', 'pkr', 'pln', 'rub', 'sek', 'sgd', 'thb', 'try', 'twd', 'usd', 'zar']
 
+function mapPrices (tickers, currency) {
+  const mapping = {}
+
+  tickers.forEach((ticker) => {
+    mapping[ticker.symbol] = parseFloat(ticker[`price_${currency.toLowerCase()}`])
+  })
+
+  return mapping
+}
+
+function query (url, currency) {
+  currency = currency.toLowerCase()
+
+  if (CURRENCY.includes(currency)) {
+    return axios.get(`${url}?convert=${currency}`)
+      .then((response) => {
+        const { data } = response
+        if (data.error) throw new Error(data.error)
+        return mapPrices(data, currency)
+      })
+  } else {
+    return Promise.reject(new ReferenceError(`${currency} is not one of the accepted currencies!`))
+  }
+}
+
+function pick (obj, ...props) {
+  return Object.assign({}, ...props.map((prop) => ({ [prop]: obj[prop] })))
+}
+
 /**
  * Returns the price of coin in the symbol given
  * @param {string} coin - Coin name. NEO or GAS.
@@ -9,18 +38,28 @@ const CURRENCY = ['aud', 'brl', 'cad', 'chf', 'clp', 'cny', 'czk', 'dkk', 'eur',
  * @return {Promise<number>} price
  */
 export const getPrice = (coin = 'NEO', currency = 'usd') => {
-  currency = currency.toLowerCase()
-  coin = coin.toLowerCase()
-  if (CURRENCY.includes(currency)) {
-    return axios.get(`https://api.coinmarketcap.com/v1/ticker/${coin}/?convert=${currency}`)
-      .then((res) => {
-        const data = res.data
-        if (data.error) throw new Error(data.error)
-        const price = data[0][`price_${currency.toLowerCase()}`]
-        if (price) return parseFloat(price)
-        else throw new Error(`Something went wrong with the CoinMarketCap API!`)
-      })
-  } else {
-    return Promise.reject(new ReferenceError(`${currency} is not one of the accepted currencies!`))
-  }
+  return query(`https://api.coinmarketcap.com/v1/ticker/${coin.toLowerCase()}/`, currency)
+    .then((mapping) => {
+      const price = mapping[coin.toUpperCase()]
+
+      if (price) return price
+      else throw new Error('Something went wrong with the CoinMarketCap API!')
+    })
+}
+
+/**
+ * Returns a mapping of the symbol for a coin to its price
+ * @param {Array<string>} coins - Coin names. NEO or GAS.
+ * @param {string} currency - Three letter currency symbol.
+ * @return {Promise<object>} object mapping symbol to price
+ */
+export const getPrices = (coins = ['NEO'], currency = 'usd') => {
+  return query(`https://api.coinmarketcap.com/v1/ticker/`, currency)
+    .then((mapping) => {
+      coins = coins.map((coin) => coin.toUpperCase())
+      const prices = pick(mapping, ...coins)
+
+      if (!coins.some((coin) => !prices[coin])) return prices
+      else throw new Error('id not found')
+    })
 }