+++ src/api/nep5.js
@@ -1,15 +1,23 @@
 import { ScriptBuilder } from '../sc'
 import { getScriptHashFromAddress, Account } from '../wallet'
 import { Query, VMZip } from '../rpc'
-import { reverseHex, fixed82num, hexstring2str } from '../utils'
+import { reverseHex, hexstring2str } from '../utils'
 import { getRPCEndpoint, getBalance } from './neonDB'
 import { Transaction } from '../transactions'
 import { ASSET_ID } from '../consts'
 
-const parseTokenInfo = VMZip(hexstring2str, hexstring2str, parseInt, fixed82num)
+/**
+ * Parses the VM output for decimals. The VM returns an integer for most cases but it can be an empty string for zero.
+ */
+const parseDecimals = (VMOutput) => {
+  if (VMOutput === '') return 0
+  return parseInt(VMOutput, 10)
+}
+const parseHexNum = (hex) => hex ? parseInt(reverseHex(hex), 16) : 0
 
-const parseTokenInfoAndBalance = VMZip(hexstring2str, hexstring2str, parseInt, fixed82num, fixed82num)
+const parseTokenInfo = VMZip(hexstring2str, hexstring2str, parseDecimals, parseHexNum)
 
+const parseTokenInfoAndBalance = VMZip(hexstring2str, hexstring2str, parseDecimals, parseHexNum, parseHexNum)
 /**
  * Queries for NEP5 Token information.
  * @param {string} url - URL of the NEO node to query.
@@ -30,7 +38,7 @@ export const getTokenInfo = (url, scriptHash) => {
         name: res[0],
         symbol: res[1],
         decimals: res[2],
-        totalSupply: res[3]
+        totalSupply: res[3] / Math.pow(10, res[2])
       }
     })
 }
@@ -45,11 +53,16 @@ export const getTokenInfo = (url, scriptHash) => {
 export const getTokenBalance = (url, scriptHash, address) => {
   const addrScriptHash = reverseHex(getScriptHashFromAddress(address))
   const sb = new ScriptBuilder()
-  const script = sb.emitAppCall(scriptHash, 'balanceOf', [addrScriptHash]).str
+  const script =
+    sb
+      .emitAppCall(scriptHash, 'decimals')
+      .emitAppCall(scriptHash, 'balanceOf', [addrScriptHash])
+      .str
   return Query.invokeScript(script, false).execute(url)
     .then((res) => {
       try {
-        return fixed82num(res.result.stack[0].value)
+        const decimals = parseDecimals(res.result.stack[0].value)
+        return parseHexNum(res.result.stack[1].value) / Math.pow(10, decimals)
       } catch (error) {
         return 0
       }
@@ -82,8 +95,8 @@ export const getToken = (url, scriptHash, address) => {
         name: res[0],
         symbol: res[1],
         decimals: res[2],
-        totalSupply: res[3],
-        balance: res.length === 5 ? res[4] : null
+        totalSupply: res[3] / Math.pow(10, res[2]),
+        balance: res.length === 5 ? res[4] / Math.pow(10, res[2]) : null
       }
     })
 }
@@ -110,10 +123,10 @@ export const doTransferToken = (net, scriptHash, fromWif, toAddress, transferAmo
       endpt = values[0]
       const balances = values[1]
       const fromAddrScriptHash = getScriptHashFromAddress(account.address)
+      const toAddrScriptHash = reverseHex(getScriptHashFromAddress(toAddress))
       const intents = [
         { assetId: ASSET_ID.GAS, value: 0.00000001, scriptHash: fromAddrScriptHash }
       ]
-      const toAddrScriptHash = reverseHex(getScriptHashFromAddress(toAddress))
       const invoke = { scriptHash, operation: 'transfer', args: [reverseHex(fromAddrScriptHash), toAddrScriptHash, transferAmount] }
       const unsignedTx = Transaction.createInvocationTx(balances, intents, invoke, gasCost, { version: 1 })
       if (signingFunction) {