+++ tests/integration/api/core.js
@@ -1,19 +1,38 @@
 import * as core from '../../../src/api/core'
+import { CONTRACTS, NEO_NETWORK, TEST_NXT_ADDRESS } from '../../../src/consts'
+import { ContractParam } from '../../../src/sc'
+import testKeys from '../../unit/testKeys.json'
 
 describe('Integration: API Core', function () {
-  this.timeout(20000)
+  this.timeout(30000)
   let mock
 
+  const useNeonDB = () => {
+    core.setApiSwitch(0)
+    mock = setupMock()
+    mock.onGet(/neoscan/).timeout()
+    mock.onAny().passThrough()
+  }
+
+  const useNeoscan = () => {
+    core.setApiSwitch(1)
+    mock = setupMock()
+    mock.onGet(/testnet-api.wallet/).timeout()
+    mock.onAny().passThrough()
+  }
   afterEach(() => {
+    core.setApiSwitch(0)
     if (mock) mock.restore()
   })
   describe('sendAsset', function () {
     it('NeonDB', () => {
-      const intent1 = core.makeIntent({ NEO: 1 }, 'ALq7AWrhAueN6mJNqk6FHJjnsEoPRytLdW')
+      useNeonDB()
+
+      const intent1 = core.makeIntent({ NEO: 1 }, testKeys.a.address)
       const config1 = {
-        net: 'TestNet',
-        address: 'ALfnhLg7rUyL6Jr98bzzoxz5J7m64fbR4s',
-        privateKey: '9ab7e154840daca3a2efadaf0df93cd3a5b51768c632f5433f86909d9b994a69',
+        net: NEO_NETWORK.TEST,
+        address: testKeys.b.address,
+        privateKey: testKeys.b.privateKey,
         intents: intent1
       }
 
@@ -25,15 +44,13 @@ describe('Integration: API Core', function () {
     })
 
     it('Neoscan', () => {
-      mock = setupMock()
-      mock.onGet(/testnet-api.wallet/).timeout()
-      mock.onAny().passThrough()
+      useNeoscan()
 
-      const intent2 = core.makeIntent({ NEO: 1 }, 'ALfnhLg7rUyL6Jr98bzzoxz5J7m64fbR4s')
+      const intent2 = core.makeIntent({ NEO: 1 }, testKeys.b.address)
       const config2 = {
-        net: 'TestNet',
-        address: 'ALq7AWrhAueN6mJNqk6FHJjnsEoPRytLdW',
-        privateKey: '7d128a6d096f0c14c3a25a2b0c41cf79661bfcb4a8cc95aaaea28bde4d732344',
+        net: NEO_NETWORK.TEST,
+        address: testKeys.a.address,
+        privateKey: testKeys.a.privateKey,
         intents: intent2
       }
       return core.sendAsset(config2)
@@ -46,10 +63,12 @@ describe('Integration: API Core', function () {
 
   describe('claimGas', function () {
     it('neonDB', () => {
+      useNeonDB()
+
       const config = {
-        net: 'TestNet',
-        address: 'ALq7AWrhAueN6mJNqk6FHJjnsEoPRytLdW',
-        privateKey: '7d128a6d096f0c14c3a25a2b0c41cf79661bfcb4a8cc95aaaea28bde4d732344'
+        net: NEO_NETWORK.TEST,
+        address: testKeys.a.address,
+        privateKey: testKeys.a.privateKey
       }
       return core.claimGas(config)
         .then((c) => {
@@ -59,14 +78,12 @@ describe('Integration: API Core', function () {
     })
 
     it('neoscan', () => {
-      mock = setupMock()
-      mock.onGet(/testnet-api.wallet/).timeout()
-      mock.onAny().passThrough()
+      useNeoscan()
 
       const config2 = {
-        net: 'TestNet',
-        address: 'ALfnhLg7rUyL6Jr98bzzoxz5J7m64fbR4s',
-        privateKey: '9ab7e154840daca3a2efadaf0df93cd3a5b51768c632f5433f86909d9b994a69'
+        net: NEO_NETWORK.TEST,
+        address: testKeys.b.address,
+        privateKey: testKeys.b.privateKey
       }
       return core.claimGas(config2)
         .then((c) => {
@@ -78,11 +95,13 @@ describe('Integration: API Core', function () {
 
   describe('doInvoke', function () {
     it('neonDB', () => {
+      useNeonDB()
+
       const config = {
-        net: 'TestNet',
-        address: 'ALq7AWrhAueN6mJNqk6FHJjnsEoPRytLdW',
-        privateKey: '7d128a6d096f0c14c3a25a2b0c41cf79661bfcb4a8cc95aaaea28bde4d732344',
-        intents: core.makeIntent({ GAS: 0.1 }, 'ALq7AWrhAueN6mJNqk6FHJjnsEoPRytLdW'),
+        net: NEO_NETWORK.TEST,
+        address: testKeys.a.address,
+        privateKey: testKeys.a.privateKey,
+        intents: core.makeIntent({ GAS: 0.1 }, testKeys.a.address),
         script: '00c1046e616d65675f0e5a86edd8e1f62b68d2b3f7c0a761fc5a67dc',
         gas: 0
       }
@@ -93,17 +112,46 @@ describe('Integration: API Core', function () {
         })
     })
 
-    it('neoscan', () => {
-      mock = setupMock()
-      mock.onGet(/testnet-api.wallet/).timeout()
-      mock.onAny().passThrough()
-
+    it('neoscan: transfer tokens', () => {
+      // This does a transferToken
+      useNeoscan()
+      const fromAddrScriptHash = ContractParam.byteArray(testKeys.b.address, 'address')
+      const toAddrScriptHash = ContractParam.byteArray(testKeys.c.address, 'address')
+      const transferAmount = ContractParam.byteArray(0.00000001, 'fixed8')
+      const script = {
+        scriptHash: CONTRACTS.TEST_LWTF,
+        operation: 'transfer',
+        args: ContractParam.array(fromAddrScriptHash, toAddrScriptHash, transferAmount)
+      }
       const config2 = {
-        net: 'TestNet',
-        address: 'ALfnhLg7rUyL6Jr98bzzoxz5J7m64fbR4s',
-        privateKey: '9ab7e154840daca3a2efadaf0df93cd3a5b51768c632f5433f86909d9b994a69',
-        intents: core.makeIntent({ GAS: 0.1 }, 'ALfnhLg7rUyL6Jr98bzzoxz5J7m64fbR4s'),
-        script: '00c1046e616d65675f0e5a86edd8e1f62b68d2b3f7c0a761fc5a67dc',
+        net: NEO_NETWORK.TEST,
+        address: testKeys.b.address,
+        privateKey: testKeys.b.privateKey,
+        intents: core.makeIntent({ GAS: 0.1 }, testKeys.b.address),
+        script,
+        gas: 0
+      }
+      return core.doInvoke(config2)
+        .then((c) => {
+          c.response.result.should.equal(true)
+          console.log(c.response.txid)
+        })
+    })
+
+    it.skip('mints tokens', () => {
+      // This does a mint tokens call
+      useNeoscan()
+      const script = {
+        scriptHash: CONTRACTS.TEST_NXT,
+        operation: 'mintTokens',
+        args: []
+      }
+      const config2 = {
+        net: NEO_NETWORK.TEST,
+        address: testKeys.a.address,
+        privateKey: testKeys.a.privateKey,
+        intents: core.makeIntent({ NEO: 1 }, TEST_NXT_ADDRESS),
+        script,
         gas: 0
       }
       return core.doInvoke(config2)